refactor into one js file per page, plus common modules
