+++ src/gallery.html
@@ -54,6 +54,6 @@
       </div>
     </div>
     <div id="snackbar"></div>
-    <script type="module" src="./js/main.js"></script>
+    <script type="module" src="./js/gallery.js"></script>
   </body>
 </html>
