+++ src/index.html
@@ -67,6 +67,6 @@
       </div>
     </div>
     <div id="snackbar"></div>
-    <script type="module" src="./js/main.js"></script>
+    <script type="module" src="./js/upload.js"></script>
   </body>
 </html>
