+++ src/settings.html
@@ -73,6 +73,6 @@
       </div>
     </div>
     <div id="snackbar"></div>
-    <script type="module" src="./js/main.js"></script>
+    <script type="module" src="./js/settings.js"></script>
   </body>
 </html>
